+++ packages/excalidraw/appState.ts
@@ -77,6 +77,7 @@ export const getDefaultAppState = (): Omit<
     isBindingEnabled: true,
     bindingPreference: "enabled",
     isMidpointSnappingEnabled: true,
+    zoomWithScrollWheel: false,
     defaultSidebarDockedPreference: false,
     isLoading: false,
     isResizing: false,
@@ -223,6 +224,7 @@ const APP_STATE_STORAGE_CONF = (<
   boxSelectionMode: { browser: true, export: false, server: false },
   bindingPreference: { browser: true, export: false, server: false },
   isMidpointSnappingEnabled: { browser: true, export: false, server: false },
+  zoomWithScrollWheel: { browser: true, export: false, server: false },
   defaultSidebarDockedPreference: {
     browser: true,
     export: false,